+++ src/i18n/workspaces.cs.json
@@ -4,8 +4,8 @@
   "showAllWorkspaces": "Zobrazit všechny pracovní prostory",
   "label": "Název",
   "owner": "Vlastník",
-  "lastEditor": "Poslední změna",
-  "lastModified": "Naposledy upraveno",
+  "lastEditor": "Spolupracovali",
+  "lastModified": "Upraveno",
   "addWorkspace": "Vytvořit pracovní prostor",
   "addingWorkspace": "Vytvářím pracovní prostor...",
   "addWorkspaceSuccess": "Pracovní prostor vytvořen",
