+++ src/i18n/workspaces.en.json
@@ -4,8 +4,8 @@
   "showAllWorkspaces": "Show all workspaces",
   "label": "Name",
   "owner": "Owner",
-  "lastEditor": "Last editor",
-  "lastModified": "Last modified",
+  "lastEditor": "Edited by",
+  "lastModified": "Modified",
   "addWorkspace": "Create workspace",
   "addingWorkspace": "Creating workspace...",
   "addWorkspaceSuccess": "Workspace was created",
